ui changes and fixes
