--- app/static/js/minorStudents.js
@@ -1,3 +1,0 @@
-$(document).ready(function(){
-    
-})
